Move `Config` type outside `src` directory.
