+++ test/invalid-data.test.ts
@@ -1,15 +1,11 @@
 import { assert } from "chai";
 import { resolve } from "path";
-import { createFormatter } from "../factory/formatter";
-import { createParser } from "../factory/parser";
-import { createProgram } from "../factory/program";
-import { Config } from "../src/Config";
-import { SchemaGenerator } from "../src/SchemaGenerator";
+import { createGenerator } from "../factory/generator";
 
 function assertSchema(name: string, type: string, message: string): void {
     it(name, () => {
         const basePath = "test/invalid-data";
-        const config: Config = {
+        const generator = createGenerator({
             path: resolve(`${basePath}/${name}/*.ts`),
             type: type,
 
@@ -17,14 +13,7 @@ function assertSchema(name: string, type: string, message: string): void {
             topRef: true,
             jsDoc: "none",
             sortProps: true,
-        };
-
-        const program = createProgram(config);
-        const generator = new SchemaGenerator(
-            program,
-            createParser(program, config),
-            createFormatter(config),
-        );
+        });
 
         assert.throws(() => generator.createSchema(type), message);
     });
